Create sample and package directories, drop unused scripts
