--- app.json
@@ -1,4 +1,0 @@
-{
-  "name": "ReactNative",
-  "displayName": "ReactNative"
-}
