+++ index.js
@@ -21,40 +21,8 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-import React from 'react';
+import SampleApp from './sample';
+import {name} from './sample/app.json';
 import {AppRegistry} from 'react-native';
-import {NavigationContainer} from '@react-navigation/native';
-import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import {ApolloClient, InMemoryCache, ApolloProvider} from '@apollo/client';
-import {STOREFRONT_DOMAIN, STOREFRONT_ACCESS_TOKEN} from '@env';
 
-import {name} from './app.json';
-
-import CatalogScreen from './src/screens/CatalogScreen';
-import SettingsScreen from './src/screens/SettingsScreen';
-
-const Tab = createBottomTabNavigator();
-
-const client = new ApolloClient({
-  uri: `https://${STOREFRONT_DOMAIN}/api/2023-10/graphql.json`,
-  cache: new InMemoryCache(),
-  headers: {
-    'Content-Type': 'application/json',
-    'X-Shopify-Storefront-Access-Token': STOREFRONT_ACCESS_TOKEN,
-  },
-});
-
-function AppWithNavigation() {
-  return (
-    <ApolloProvider client={client}>
-      <NavigationContainer>
-        <Tab.Navigator>
-          <Tab.Screen name="Catalog" component={CatalogScreen} />
-          <Tab.Screen name="Settings" component={SettingsScreen} />
-        </Tab.Navigator>
-      </NavigationContainer>
-    </ApolloProvider>
-  );
-}
-
-AppRegistry.registerComponent(name, () => AppWithNavigation);
+AppRegistry.registerComponent(name, () => SampleApp);
