+++ sample/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "ReactNative",
+  "displayName": "Shopify Checkout"
+}
