+++ sample/src/screens/SettingsScreen.tsx
