--- scripts/test_package
@@ -1,12 +1,0 @@
- #!/usr/bin/env bash
-
-set -ex
-set -eo pipefail
-
-if [[ -n $CURRENT_SIMULATOR_UUID ]]; then
-    dest="id=$CURRENT_SIMULATOR_UUID"
-else
-    dest="platform=iOS Simulator,name=iPhone 14"
-fi
-
-xcodebuild test -scheme ShopifyCheckout -sdk iphonesimulator -destination "$dest" -skipPackagePluginValidation | xcpretty -c
